Fix backup validation logic
