+++ packages/server/src/utils/backups/s3-storage-class.ts
@@ -1,3 +1,6 @@
+import { findDestinationById } from "@dokploy/server/services/destination";
+import { TRPCError } from "@trpc/server";
+
 const PROVIDER_STORAGE_CLASS_OPTIONS = {
 	AWS: [
 		"STANDARD",
@@ -57,4 +60,35 @@ export const isValidS3StorageClassForProvider = ({
 	}
 
 	return providerStorageClasses.includes(normalizedStorageClass);
+};
+
+export const validateS3StorageClassForDestination = async ({
+	destinationId,
+	storageClass,
+}: {
+	destinationId: string;
+	storageClass?: string | null;
+}) => {
+	const normalizedStorageClass = normalizeS3StorageClass(storageClass);
+	if (!normalizedStorageClass) {
+		return;
+	}
+
+	const destination = await findDestinationById(destinationId);
+	const provider = destination.provider;
+	const supportedStorageClasses = getS3StorageClassesForProvider(provider);
+
+	if (supportedStorageClasses.length === 0) {
+		throw new TRPCError({
+			code: "BAD_REQUEST",
+			message: `Storage class is not supported for provider "${provider || "Unknown"}".`,
+		});
+	}
+
+	if (!supportedStorageClasses.includes(normalizedStorageClass)) {
+		throw new TRPCError({
+			code: "BAD_REQUEST",
+			message: `Invalid storage class for provider "${provider}". Allowed values: ${supportedStorageClasses.join(", ")}.`,
+		});
+	}
 };
